Minor fixes and code editing

+++ KadLookupGraph.cpp
@@ -556,7 +556,7 @@ void CKadLookupGraph::UpdateToolTip()
 		if (sEntry->m_liReceivedFromIdx.IsEmpty())
 			strDiscovered = GetResString(IDS_ROUTINGTABLE);
 		else
-			strDiscovered.Format(_T("%i %s"), sEntry->m_liReceivedFromIdx.GetCount(), (LPCTSTR)GetResString(IDS_NODES));
+			strDiscovered.Format(_T("%i %s"), (int)sEntry->m_liReceivedFromIdx.GetCount(), (LPCTSTR)GetResString(IDS_NODES));
 
 		CString strFoundNodes;
 		if (sEntry->m_dwAskedContactsTime > 0)
